Adding csv for hate_crimes
This file is trimmed and altered so that there is a column for each targeted group. Some crimes were against multiple groups. This is my attempt to work with the data and hopefully be able to track by targeted group by making each a column and using a script to denote an active targeted group for each crime.